fix: export float 32 for np arrays

+++ dquartic/utils/data_generation.py
@@ -102,13 +102,13 @@ def create_parquet_data(input_file: str, current_iso, slices_ms1, slices_ms2, wi
             'mz_end': current_iso['mzEnd'],
             'rt_start': window[0],
             'rt_end': window[-1],
-            'ms1_data': slice_ms1.flatten(),
-            'ms2_data': slice_ms2.flatten(),
+            'ms1_data': slice_ms1.flatten().astype(np.float32),
+            'ms2_data': slice_ms2.flatten().astype(np.float32),
             'ms1_shape': slice_ms1.shape,
             'ms2_shape': slice_ms2.shape,
-            'rt_values': np.array(window),
-            'mz_values_ms1': unique_mz.to_numpy(),
-            'mz_values_ms2': unique_mz_ms2.to_numpy()
+            'rt_values': np.array(window).astype(np.float32),
+            'mz_values_ms1': unique_mz.to_numpy().astype(np.float32),
+            'mz_values_ms2': unique_mz_ms2.to_numpy().astype(np.float32),
         }
         data.append(slice_data)
